9 homework 3 sprint is done

+++ src/services/sessionService.ts
@@ -1,18 +1,26 @@
 import {jwtService} from "./jwtService";
 import {sessionRepository} from "../repositories/sessionRepository";
+import {sessionQueryRepo} from "../repositories/sessionQueryRepo";
+import {DomainExceptions} from "../helpers/DomainExceptions";
 
 export const sessionsService = {
     deleteAllSessionsExceptActive: async (refreshToken: string) => {
         const payload = await jwtService.checkToken(refreshToken);
         const userId = payload.userId
         const deviceId = payload.deviceId
-        console.log("del sessionAll-->",userId, deviceId)
         return await sessionRepository.deleteAllSessionsExceptActive(userId, deviceId)
     },
     deleteSessionByDeviceId: async (refreshToken: string, deviceId: string) => {
-        const payload = await jwtService.checkToken(refreshToken);
-        const userId = payload.userId
-        console.log("del sessionById-->",userId, deviceId)
-        return await sessionRepository.deleteSessionByDeviceId(userId, deviceId)
+        // try {
+            const payload = await jwtService.checkToken(refreshToken);
+            const userId = payload.userId
+            if(!userId){
+                throw new DomainExceptions(401, 'unauthorized:invalid token')
+            }
+            return await sessionRepository.deleteSessionByDeviceId(userId, deviceId)
+        // } catch (error) {
+        //     console.log(error)
+        //     return false
+        // }
     }
 }
